+++ src/wetterdienst/provider/chmi/observation/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (C) 2018-2025, earthobservations developers.
+# Distributed under the MIT License. See LICENSE for more info.
+"""CHMI (Czech Hydrometeorological Institute) observation provider."""
+
+from wetterdienst.provider.chmi.observation.api import ChmiObservationRequest
+from wetterdienst.provider.chmi.observation.metadata import ChmiObservationMetadata
+
+__all__ = ["ChmiObservationMetadata", "ChmiObservationRequest"]